+++ src/__generated__/FollowArtistButton_artist.graphql.ts
@@ -6,6 +6,7 @@ export type FollowArtistButton_artist$ref = typeof _FollowArtistButton_artist$re
 export type FollowArtistButton_artist = {
     readonly id: string;
     readonly slug: string;
+    readonly internalID: string;
     readonly is_followed: boolean | null;
     readonly " $refType": FollowArtistButton_artist$ref;
 };
@@ -36,11 +37,18 @@ const node: ReaderFragment = {
     {
       "kind": "ScalarField",
       "alias": null,
+      "name": "internalID",
+      "args": null,
+      "storageKey": null
+    },
+    {
+      "kind": "ScalarField",
+      "alias": null,
       "name": "is_followed",
       "args": null,
       "storageKey": null
     }
   ]
 };
-(node as any).hash = '7a627060bddd826f654a02c63425e824';
+(node as any).hash = 'eee7f655b41839900361749e24f62a7e';
 export default node;