+++ src/lib/Components/Inbox/Conversations/Composer.tsx
@@ -115,7 +115,7 @@ export default class Composer extends React.Component<Props, State> {
             autoFocus={typeof jest === "undefined" /* TODO: https://github.com/facebook/jest/issues/3707 */}
           />
           <TouchableWithoutFeedback disabled={disableSendButton} onPress={this.submitText.bind(this)}>
-            <SendButton disabled={disableSendButton}>SEND</SendButton>
+            <SendButton disabled={disableSendButton}>Send</SendButton>
           </TouchableWithoutFeedback>
         </Container>
       </StyledKeyboardAvoidingView>