+++ src/lib/Scenes/Artwork/Components/ArtworkHistory.tsx
@@ -1,6 +1,7 @@
 import { Box, Join, Sans, Spacer } from "@artsy/palette"
 import { ArtworkHistory_artwork } from "__generated__/ArtworkHistory_artwork.graphql"
 import { ReadMore } from "lib/Components/ReadMore"
+import { Schema } from "lib/utils/track"
 import React from "react"
 import { createFragmentContainer, graphql } from "react-relay"
 
@@ -12,21 +13,30 @@ export class ArtworkHistory extends React.Component<ArtworkHistoryProps> {
   render() {
     const { provenance, exhibition_history, literature } = this.props.artwork
     const sections = [
-      { title: "Provenance", value: provenance },
-      { title: "Exhibition History", value: exhibition_history },
-      { title: "Bibliography", value: literature },
+      { title: "Provenance", value: provenance, contextModule: Schema.ContextModules.Provenance },
+      {
+        title: "Exhibition History",
+        value: exhibition_history,
+        contextModule: Schema.ContextModules.ExhibitionHistory,
+      },
+      { title: "Bibliography", value: literature, contextModule: Schema.ContextModules.Bibliography },
     ]
 
     const displaySections = sections.filter(i => i.value != null)
 
     return (
       <Join separator={<Spacer pb={3} />}>
-        {displaySections.map(({ title, value }, index) => (
+        {displaySections.map(({ title, value, contextModule }, index) => (
           <Box key={index}>
             <Sans size="3" weight="medium" pb={2}>
               {title}
             </Sans>
-            <ReadMore content={value} maxChars={140} />
+            <ReadMore
+              content={value}
+              maxChars={140}
+              trackingFlow={Schema.Flow.ArtworkDetails}
+              contextModule={contextModule}
+            />
           </Box>
         ))}
       </Join>