+++ src/lib/Scenes/Artwork/Components/FollowArtistButton.tsx
@@ -1,6 +1,7 @@
 import { Sans } from "@artsy/palette"
 import { FollowArtistButton_artist } from "__generated__/FollowArtistButton_artist.graphql"
 import { FollowArtistButtonMutation } from "__generated__/FollowArtistButtonMutation.graphql"
+import { Schema, track } from "lib/utils/track"
 import React from "react"
 import { TouchableWithoutFeedback } from "react-native"
 import { commitMutation, createFragmentContainer, graphql, RelayProp } from "react-relay"
@@ -8,10 +9,22 @@ import { commitMutation, createFragmentContainer, graphql, RelayProp } from "rea
 interface Props {
   artist: FollowArtistButton_artist
   relay: RelayProp
+  contextModule?: string
 }
 
+@track()
 export class FollowArtistButton extends React.Component<Props> {
-  handleFollowArtist = () => {
+  @track((props: Props) => {
+    return {
+      action_name: props.artist.is_followed ? Schema.ActionNames.ArtistFollow : Schema.ActionNames.ArtistUnfollow,
+      action_type: Schema.ActionTypes.Success,
+      owner_id: props.artist.internalID,
+      owner_slug: props.artist.slug,
+      owner_type: Schema.OwnerEntityTypes.Artist,
+      context_module: props.contextModule ? props.contextModule : null,
+    } as any
+  })
+  handleFollowArtist() {
     const { artist, relay } = this.props
     commitMutation<FollowArtistButtonMutation>(relay.environment, {
       mutation: graphql`
@@ -45,7 +58,7 @@ export class FollowArtistButton extends React.Component<Props> {
   render() {
     const followButtonText = this.props.artist.is_followed ? "Following" : "Follow"
     return (
-      <TouchableWithoutFeedback onPress={this.handleFollowArtist.bind(this)}>
+      <TouchableWithoutFeedback onPress={() => this.handleFollowArtist()}>
         <Sans color="black60" weight="medium" size="3t">
           {followButtonText}
         </Sans>
@@ -59,6 +72,7 @@ export const FollowArtistButtonFragmentContainer = createFragmentContainer(Follo
     fragment FollowArtistButton_artist on Artist {
       id
       slug
+      internalID
       is_followed
     }
   `,